Updated SDL library sources to latest, resolving ALT-TAB and Com_EventQueue overflow issue on windows

+++ src/libs/libsdl/src/video/Xext/Xxf86dga/XF86DGA2.c
@@ -960,8 +960,8 @@ DGAMapPhysical(
 #else
 	return False;
 #endif
-    pMap->virtual = mmap(NULL, size, PROT_READ | PROT_WRITE, 
-			MAP_FILE | MAP_SHARED, pMap->fd, (off_t)base);
+    pMap->virtual = mmap(NULL, size, PROT_READ | PROT_WRITE,
+			MAP_FILE | MAP_SHARED, pMap->fd, (off_t)((size_t)base));
     if (pMap->virtual == (void *)-1)
 	return False;
 #endif
